[Linux] Enable clDNN backend on Linux
Remove linux temporary impl.
