+++ services/ml/BUILD.gn
@@ -16,15 +16,19 @@ source_set("lib") {
 
   if (is_linux) {
     sources += [
-      "compilation_impl_linux.cc",
-      "compilation_impl_linux.h",
-      "execution_impl_linux.cc",
-      "execution_impl_linux.h",
-      "model_impl_linux.cc",
-      "model_impl_linux.h",
-      "neural_network_impl_linux.cc",
-      "neural_network_impl_linux.h",
+      "compilation_impl_cl_dnn.cc",
+      "compilation_impl_cl_dnn.h",
+      "execution_impl_cl_dnn.cc",
+      "execution_impl_cl_dnn.h",
+      "model_impl_cl_dnn.cc",
+      "model_impl_cl_dnn.h",
+      "neural_network_impl_cl_dnn.cc",
+      "neural_network_impl_cl_dnn.h",
     ]
+
+    lib_dirs = [ "//third_party/clDNN/build/out/Linux64/Release" ]
+
+    libs = [ "clDNN64" ]
   } else if (is_android) {
     sources += [
       "compilation_impl_android.cc",
@@ -81,14 +85,14 @@ source_set("lib") {
     ]
   } else if (is_win) {
     sources += [
-      "compilation_impl_win.cc",
-      "compilation_impl_win.h",
-      "execution_impl_win.cc",
-      "execution_impl_win.h",
-      "model_impl_win.cc",
-      "model_impl_win.h",
-      "neural_network_impl_win.cc",
-      "neural_network_impl_win.h",
+      "compilation_impl_cl_dnn.cc",
+      "compilation_impl_cl_dnn.h",
+      "execution_impl_cl_dnn.cc",
+      "execution_impl_cl_dnn.h",
+      "model_impl_cl_dnn.cc",
+      "model_impl_cl_dnn.h",
+      "neural_network_impl_cl_dnn.cc",
+      "neural_network_impl_cl_dnn.h",
     ]
 
     lib_dirs = [ "//third_party/clDNN/build/out/Windows64/Release" ]
@@ -109,6 +113,10 @@ source_set("lib") {
     deps += [ ":copy_cldnn_dll" ]
   }
 
+  if (is_linux) {
+    deps += [ ":copy_cldnn_so" ]
+  }
+
   public_deps = [
     "//base",
     "//services/ml/public/interfaces",
@@ -121,6 +129,18 @@ if (is_win) {
   copy("copy_cldnn_dll") {
     sources = [
       "$dll_path/clDNN64.dll",
+    ]
+    outputs = [
+      "$root_out_dir/{{source_file_part}}",
+    ]
+  }
+}
+
+if (is_linux) {
+  so_path = "//third_party/clDNN/build/out/Linux64/Release"
+  copy("copy_cldnn_so") {
+    sources = [
+      "$so_path/libclDNN64.so",
     ]
     outputs = [
       "$root_out_dir/{{source_file_part}}",
